requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
asyncpg>=0.29.0
aiomysql>=0.2.0
gitpython>=3.1.40
Pillow>=10.1.0  # For image generation (optional)
aiofiles>=23.2.0